requirements.txt
beanie==1.26.0
fastapi==0.112.0
motor==3.5.1
passlib==1.7.4
pydantic==2.8.2
pydantic_settings==2.4.0
PyJWT==2.9.0
PyJWT==2.9.0
pymongo==4.8.0
